Send Icinga Reports Asynchronously
It's an anti-pattern to call a backend service like Icinga synchronously,
because that leads to long chains of calls were the whole chain breaks
if some service in between is slow to respond.

We use an executor here to execute the monitoring calls asynchronously.
It's also important that the executor has no unbounded queue, because
that would lead to filling the memory with tasks that will be never
executed. Instead we use a relatively small queue of 100 tasks and a
policy to abort the oldest waiting tasks. So with a slow responding
Icinga, we will have holes in monitoring reports but always try to
deliver the 100 newest reports.